Sanitize wrapper identifiers
