+++ crates/rstest-bdd-macros/src/codegen/wrapper/emit.rs
@@ -2,6 +2,7 @@
 
 use super::args::{ArgumentCollections, CallArg, DataTableArg, DocStringArg, FixtureArg, StepArg};
 use crate::codegen::keyword_to_token;
+use crate::utils::ident::sanitize_ident;
 use proc_macro2::TokenStream as TokenStream2;
 use quote::{format_ident, quote};
 use std::sync::atomic::{AtomicUsize, Ordering};
@@ -120,6 +121,10 @@ static COUNTER: AtomicUsize = AtomicUsize::new(0);
 
 /// Generate unique identifiers for the wrapper components.
 ///
+/// The provided step function identifier may contain Unicode. It is
+/// sanitised to ASCII characters before constructing constant names to avoid
+/// emitting invalid identifiers.
+///
 /// Returns identifiers for the wrapper function, fixture array constant, and
 /// pattern constant.
 fn generate_wrapper_identifiers(
@@ -127,7 +132,7 @@ fn generate_wrapper_identifiers(
     id: usize,
 ) -> (proc_macro2::Ident, proc_macro2::Ident, proc_macro2::Ident) {
     let wrapper_ident = format_ident!("__rstest_bdd_wrapper_{}_{}", ident, id);
-    let ident_upper = ident.to_string().to_uppercase();
+    let ident_upper = sanitize_ident(&ident.to_string()).to_ascii_uppercase();
     let const_ident = format_ident!("__RSTEST_BDD_FIXTURES_{}_{}", ident_upper, id);
     let pattern_ident = format_ident!("__RSTEST_BDD_PATTERN_{}_{}", ident_upper, id);
     (wrapper_ident, const_ident, pattern_ident)
@@ -345,5 +350,29 @@ pub(crate) fn generate_wrapper_code(config: &WrapperConfig<'_>) -> TokenStream2 
     quote! {
         #body
         #registration
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::generate_wrapper_identifiers;
+    use quote::format_ident;
+    use syn::parse_str;
+
+    #[test]
+    fn sanitizes_constant_identifiers() {
+        let ident = match parse_str::<syn::Ident>("préférence") {
+            Ok(i) => i,
+            Err(e) => panic!("parse identifier: {e}"),
+        };
+        let (_, const_ident, pattern_ident) = generate_wrapper_identifiers(&ident, 3);
+        assert_eq!(
+            const_ident,
+            format_ident!("__RSTEST_BDD_FIXTURES_PR_F_RENCE_3")
+        );
+        assert_eq!(
+            pattern_ident,
+            format_ident!("__RSTEST_BDD_PATTERN_PR_F_RENCE_3")
+        );
     }
 }
