+++ crates/rstest-bdd-macros/src/macros/scenarios.rs
@@ -9,31 +9,8 @@ use std::path::{Path, PathBuf};
 use crate::codegen::scenario::{ScenarioConfig, generate_scenario_code};
 use crate::parsing::feature::{extract_scenario_steps, parse_and_load_feature};
 use crate::utils::errors::error_to_tokens;
+use crate::utils::ident::sanitize_ident;
 use gherkin::Feature;
-
-/// Sanitize a string so it may be used as a Rust identifier.
-///
-/// Only ASCII alphanumeric characters are preserved; all other characters
-/// (including Unicode) are replaced with underscores. The result is lowercased.
-/// Identifiers starting with a digit gain a leading underscore.
-///
-/// Note: Unicode characters are not supported and will be replaced with
-/// underscores.
-/// TODO: Consider supporting Unicode normalization in the future.
-fn sanitize_ident(input: &str) -> String {
-    let mut ident = String::new();
-    for c in input.chars() {
-        if c.is_ascii_alphanumeric() {
-            ident.push(c.to_ascii_lowercase());
-        } else {
-            ident.push('_');
-        }
-    }
-    if ident.is_empty() || matches!(ident.chars().next(), Some(c) if c.is_ascii_digit()) {
-        ident.insert(0, '_');
-    }
-    ident
-}
 
 /// Recursively collect all `.feature` files under `base`.
 fn collect_feature_files(base: &Path) -> std::io::Result<Vec<PathBuf>> {
@@ -263,28 +240,8 @@ pub(crate) fn scenarios(input: TokenStream) -> TokenStream {
 
 #[cfg(test)]
 mod tests {
-    use super::{dedupe_name, sanitize_ident};
+    use super::dedupe_name;
     use std::collections::HashSet;
-
-    #[test]
-    fn sanitizes_invalid_identifiers() {
-        assert_eq!(sanitize_ident("Hello world!"), "hello_world_");
-    }
-
-    #[test]
-    fn sanitizes_leading_digit() {
-        assert_eq!(sanitize_ident("123abc"), "_123abc");
-    }
-
-    #[test]
-    fn sanitizes_empty_input() {
-        assert_eq!(sanitize_ident(""), "_");
-    }
-
-    #[test]
-    fn sanitizes_unicode() {
-        assert_eq!(sanitize_ident("Crème—brûlée"), "cr_me_br_l_e");
-    }
 
     #[test]
     fn deduplicates_duplicate_titles() {
