+++ crates/rstest-bdd-macros/src/utils/mod.rs
@@ -2,3 +2,4 @@
 
 pub(crate) mod errors;
 pub(crate) mod fixtures;
+pub(crate) mod ident;
