+++ crates/rstest-bdd-macros/tests/fixtures/step_macros_unicode.rs
@@ -1,0 +1,13 @@
+//! Step definitions with Unicode identifiers used by trybuild tests.
+use rstest_bdd_macros::{given, when, then};
+
+#[given("précondition")]
+fn précondition() {}
+
+#[when("acción")]
+fn acción() {}
+
+#[then("résultat")]
+fn résultat() {}
+
+fn main() {}
